ajout de l'api de synchronisation des producteurs

+++ resources/views/livewire/offer/offer-component.blade.php
@@ -29,12 +29,13 @@
                             <thead class="bg-amber-800 bt-table">
                                 <tr class="">
                                     <th scope="col" class="rounded-tl-lg px-6 py-4">#</th>
+                                    <th scope="col" class="px-6 py-4">Code</th>
                                     <th scope="col" class="px-6 py-4">Coopérative <br> Département</th>
                                     <th scope="col" class="px-6 py-4">Certification</th>
                                     <th scope="col" class="px-6 py-4">Poids (kg)</th>
                                     <th scope="col" class="px-6 py-4">Prix unitaire (FCFA)</th>
                                     <th scope="col" class="px-6 py-4">Prix Total (FCFA)</th>
-                                    <th scope="col" class="px-6 py-4">Conditionnement</th>
+                                    <!-- <th scope="col" class="px-6 py-4">Conditionnement</th> -->
                                     <th scope="col" class="px-6 py-4">Status</th>
                                     <th scope="col" class="rounded-tr-lg px-6 py-4">Action</th>
                                 </tr>
@@ -47,6 +48,7 @@
                                 <?php $i++ ?>
                                 <tr class="border-b border-t-2 border-amber-800 {{ $i % 2 !== 0 ? '' : 'bg-amber-100' }} dark:border-amber-900">
                                     <td class="whitespace-nowrap px-6 py-4 font-medium" wire:key="{{ $offer->id }}">{{ $i }}</td>
+                                    <td class="whitespace-nowrap px-6 py-4 font-medium">{{ $offer->code }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">{{ $offer->agribusiness->sigle ?? '-' }} <br> {{ $offer->agribusiness->departement->name ?? '-' }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">{{ $offer->certification->name }}</td>
                                     
@@ -54,15 +56,19 @@
                                     
                                     <td class="whitespace-nowrap px-6 py-4">{{ $offer->selling_price }}</td>
                                     <td class="whitespace-nowrap px-6 py-4">{{ ($offer->selling_price * $offer->weight ) }}</td>
-                                    <td class="whitespace-nowrap px-6 py-4">{{ $offer->qte }}  {{ $offer->type_packaging }}(s) </td>
+                                    <!-- <td class="whitespace-nowrap px-6 py-4">{{ $offer->type_package_id }}</td> -->
                                     <td class="whitespace-nowrap px-6 py-4">-</td>
                                     
                                     <td class="whitespace-nowrap px-6 py-4">
+                                    @if(auth()->check() && (auth()->user()->isPlateformAdmin() || auth()->user()->isAgribusinessAdmin ))
+                                        @canany(['SUPPRIMER UNE OFFRE'])
                                         <button class="btn-success" wire:click='edit("{{$offer->id}}")' >
                                             <svg width="15px" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
                                                 <path fill-rule="evenodd" clip-rule="evenodd" d="M21.1213 2.70705C19.9497 1.53548 18.0503 1.53547 16.8787 2.70705L15.1989 4.38685L7.29289 12.2928C7.16473 12.421 7.07382 12.5816 7.02986 12.7574L6.02986 16.7574C5.94466 17.0982 6.04451 17.4587 6.29289 17.707C6.54127 17.9554 6.90176 18.0553 7.24254 17.9701L11.2425 16.9701C11.4184 16.9261 11.5789 16.8352 11.7071 16.707L19.5556 8.85857L21.2929 7.12126C22.4645 5.94969 22.4645 4.05019 21.2929 2.87862L21.1213 2.70705ZM18.2929 4.12126C18.6834 3.73074 19.3166 3.73074 19.7071 4.12126L19.8787 4.29283C20.2692 4.68336 20.2692 5.31653 19.8787 5.70705L18.8622 6.72357L17.3068 5.10738L18.2929 4.12126ZM15.8923 6.52185L17.4477 8.13804L10.4888 15.097L8.37437 15.6256L8.90296 13.5112L15.8923 6.52185ZM4 7.99994C4 7.44766 4.44772 6.99994 5 6.99994H10C10.5523 6.99994 11 6.55223 11 5.99994C11 5.44766 10.5523 4.99994 10 4.99994H5C3.34315 4.99994 2 6.34309 2 7.99994V18.9999C2 20.6568 3.34315 21.9999 5 21.9999H16C17.6569 21.9999 19 20.6568 19 18.9999V13.9999C19 13.4477 18.5523 12.9999 18 12.9999C17.4477 12.9999 17 13.4477 17 13.9999V18.9999C17 19.5522 16.5523 19.9999 16 19.9999H5C4.44772 19.9999 4 19.5522 4 18.9999V7.99994Z" fill="#fff"/>
                                             </svg>
-                                        </button>
+                                        </button>                                            
+                                        @endcanany
+                                        @canany(['SUPPRIMER UNE OFFRE'])
                                         <button class="btn-danger flex flex-row" wire:click='deleteForm("{{$offer->id}}")' >
                                             <svg width="15px" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
                                                 <path d="M12 2.75C11.0215 2.75 10.1871 3.37503 9.87787 4.24993C9.73983 4.64047 9.31134 4.84517 8.9208 4.70713C8.53026 4.56909 8.32557 4.1406 8.46361 3.75007C8.97804 2.29459 10.3661 1.25 12 1.25C13.634 1.25 15.022 2.29459 15.5365 3.75007C15.6745 4.1406 15.4698 4.56909 15.0793 4.70713C14.6887 4.84517 14.2602 4.64047 14.1222 4.24993C13.813 3.37503 12.9785 2.75 12 2.75Z" fill="#fff"/>
@@ -71,8 +77,27 @@
                                                 <path d="M9.42546 10.2537C9.83762 10.2125 10.2051 10.5132 10.2464 10.9254L10.7464 15.9254C10.7876 16.3375 10.4869 16.7051 10.0747 16.7463C9.66256 16.7875 9.29502 16.4868 9.25381 16.0746L8.75381 11.0746C8.71259 10.6625 9.0133 10.2949 9.42546 10.2537Z" fill="#fff"/>
                                                 <path d="M15.2464 11.0746C15.2876 10.6625 14.9869 10.2949 14.5747 10.2537C14.1626 10.2125 13.795 10.5132 13.7538 10.9254L13.2538 15.9254C13.2126 16.3375 13.5133 16.7051 13.9255 16.7463C14.3376 16.7875 14.7051 16.4868 14.7464 16.0746L15.2464 11.0746Z" fill="#fff"/>
                                             </svg>
-                                            
                                         </button>
+                                        @endcanany
+                                    @else
+                                        <button class="btn-success"  wire:click='demande("{{ $offer->id }}")'>
+                                            <!-- <svg width="15px" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                            <style type="text/css">
+                                                    .st0{fill:#0000;}
+                                                </style>
+                                                <g>
+                                                    <path class="st0" d="M165.013,288.946h75.034c6.953,0,12.609,5.656,12.609,12.608v26.424c0,7.065,3.659,9.585,7.082,9.585
+                                                        c2.106,0,4.451-0.936,6.78-2.702l90.964-69.014c3.416-2.589,5.297-6.087,5.297-9.844c0-3.762-1.881-7.259-5.297-9.849
+                                                        l-90.964-69.014c-2.329-1.766-4.674-2.702-6.78-2.702c-3.424,0-7.082,2.519-7.082,9.584v26.425c0,6.952-5.656,12.608-12.609,12.608
+                                                        h-75.034c-8.707,0-15.79,7.085-15.79,15.788v34.313C149.223,281.862,156.305,288.946,165.013,288.946z"/>
+                                                    <path class="st0" d="M256,0C114.842,0,0.002,114.84,0.002,256S114.842,512,256,512c141.158,0,255.998-114.84,255.998-256
+                                                        S397.158,0,256,0z M256,66.785c104.334,0,189.216,84.879,189.216,189.215S360.334,445.215,256,445.215S66.783,360.336,66.783,256
+                                                        S151.667,66.785,256,66.785z"/>
+                                                </g>
+                                            </svg> -->
+                                            Commander
+                                        </button>   
+                                    @endif  
                                     </td>
